+++ cishouseholds/pipeline/high_level_transformations.py
@@ -152,26 +152,6 @@ from cishouseholds.pyspark_utils import get_or_create_spark_session
 from cishouseholds.validate_class import SparkValidate
 
 
-def generate_lab_report(df: DataFrame, current_date=F.current_timestamp()) -> DataFrame:
-    """
-    Generate lab report of latest 7 days of results
-    """
-    df = df.filter(F.date_sub(current_date, 7) < F.col("survey_completed_datetime"))
-    swab_df = df.select("swab_sample_barcode", "swab_taken_datetime", "survey_completed_datetime").filter(
-        ~(
-            ((F.col("swab_taken_datetime").isNull()) & (F.col("survey_completed_datetime").isNull()))
-            | F.col("swab_sample_barcode").isNull()
-        )
-    )
-    blood_df = df.select("blood_sample_barcode", "blood_taken_datetime", "survey_completed_datetime").filter(
-        ~(
-            ((F.col("blood_taken_datetime").isNull()) & (F.col("survey_completed_datetime").isNull()))
-            | F.col("blood_sample_barcode").isNull()
-        )
-    )
-    return swab_df, blood_df
-
-
 def transform_cis_soc_data(df: DataFrame) -> DataFrame:
     """
     transform and process cis soc data
