+++ cishouseholds/pipeline/reporting.py
@@ -13,6 +13,26 @@ from cishouseholds.pipeline.load import extract_from_table
 from cishouseholds.pipeline.load import get_run_id
 
 
+def generate_lab_report(df: DataFrame, current_date=F.current_timestamp()) -> DataFrame:
+    """
+    Generate lab report of latest 7 days of results
+    """
+    df = df.filter(F.date_sub(current_date, 7) < F.col("survey_completed_datetime"))
+    swab_df = df.select("swab_sample_barcode", "swab_taken_datetime", "survey_completed_datetime").filter(
+        ~(
+            ((F.col("swab_taken_datetime").isNull()) & (F.col("survey_completed_datetime").isNull()))
+            | F.col("swab_sample_barcode").isNull()
+        )
+    )
+    blood_df = df.select("blood_sample_barcode", "blood_taken_datetime", "survey_completed_datetime").filter(
+        ~(
+            ((F.col("blood_taken_datetime").isNull()) & (F.col("survey_completed_datetime").isNull()))
+            | F.col("blood_sample_barcode").isNull()
+        )
+    )
+    return swab_df, blood_df
+
+
 def dfs_to_bytes_excel(sheet_df_map: Dict[str, DataFrame]) -> BytesIO:
     """
     Convert a dictionary of Spark DataFrames into an Excel Object.
